leveling fixed
doesnt work with module.exports anymore, cannot figure out why for the life of me. Seems that the `messageCreate` event is called on startup for some godforsaken reason
