--- .env.example
@@ -1,3 +1,0 @@
-token=your_token
-WEATHER_API_KEY=your_weather_api_key
-MONGO_URI=your_mongo_uri
